fix: support 4 GB / 20 h audiobooks + indexer reads ID3 chapters
Two hard blockers for a 1.63 GB / 20 h audiobook, both found by inspection
before the user's test:
- data.json Audio slot size_maximum was 1 GB → the APF refuses to open a
  1.63 GB file outright. Raised to 0xFFFFFFFF (the field is 32-bit).
- track_t.dur_s was uint16_t → 65535 s = 18.2 h. A 20 h book (72000 s) wrapped,
  corrupting the duration readout, the progress bar and seek_byte_off(). Now
  uint32_t; the parser's clamp is gone. Seek math re-checked across the range
  (64-bit intermediate; 19h59 → byte 1,749,956,089, exact). fsize stays uint32
  (4 GB ceiling), which the raised slot limit now matches.

Chapters (audiobook L/R navigation, firmware side to follow): music-metadata
exposes CHAP start/end but leaves the embedded sub-frames empty, losing the
titles — so chapters.js parses the raw ID3v2 tag (id3.org chapters-1.0): CHAP
body = elementID\0 + 4 timestamps + nested frames, TIT2 = title. Handles v2.3
(plain size) and v2.4 (synchsafe), all four text encodings, and names unnamed
chapters. Indexer emits chapters:[{s,t}] per track; verified on a purpose-built
chaptered MP3 (accented titles intact) and on a chapterless file (→ []).

Also: M7a validated on hardware — SRAM boot + MP3 + a 265 MB/1h56 file play.
The SRAM move freed BRAM: stack is now 54 KB (was 3.1 KB).

+++ firmware/src/lib.h
@@ -20,7 +20,8 @@ typedef struct {
   uint32_t path_off;     // byte offset of the path string inside library.json
   uint32_t fsize;        // audio file size in bytes
   uint16_t path_len;
-  uint16_t dur_s;        // duration in seconds (0 = unknown)
+  uint32_t dur_s;        // duration in seconds (0 = unknown). 32-bit: a 20 h
+                         // audiobook is 72000 s, well past a uint16's 18.2 h.
   uint16_t kbps;         // nominal bitrate (0 = unknown)
   uint8_t  format;       // FMT_*
 } track_t;
